move a small set of non-scalar tests out of scalar.test_timestamp
